+++ app/src/main/java/com/haque/baseline/data/source/source/dto/DailyWeatherDTO.kt
@@ -6,7 +6,7 @@ data class DailyWeatherDTO (
     val time: List<String>,
 
     @field:Json(name = "weathercode")
-    val weatherCodes: List<Int>,
+    val weatherCodes: List<Long>,
 
     @field:Json(name = "temperature_2m_max")
     val maxTemperatures: List<Double>,
@@ -19,4 +19,4 @@ data class DailyWeatherDTO (
 
     @field:Json(name = "precipitation_hours")
     val precipitationHours: List<Double>
-        )
+    )
